backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
ccxt==4.1.15
pandas-ta==0.3.14b0
pandas==2.1.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0